# docs/openapi.yaml
openapi: 3.0.3
info:
  title: APP_ModularDeGestion API
  description: >
    Contrato de la API para el sistema de gestión modular multi-negocio.
    Fuente de verdad compartida entre frontend y backend — ver
    docs/ARCHITECTURE.md para el contexto de dominio.

    Auth: todo endpoint requiere `Authorization: Bearer <token>` (ver
    POST /auth/login), excepto /health y el login mismo. El negocio_id de
    cualquier ruta /negocios/{negocioId}/... se valida contra el negocio_id
    del token — no alcanza con mandar un negocio_id distinto en la URL para
    operar sobre otro negocio. Un usuario con rol admin no está atado a
    ningún negocio y puede operar sobre cualquiera; el resto de roles
    (hoy solo "dueño") solo sobre el suyo. Devuelve 401 sin token o con uno
    inválido/expirado, 403 con un token válido pero sin permiso sobre ese
    negocio_id o sin rol admin donde se requiere.
  version: 0.1.0
servers:
  - url: http://localhost:8000/api/v1
    description: Desarrollo local
...
security:
  - bearerAuth: []
paths:
...
  /negocios/{negocioId}/movimientos:
    parameters:
      - $ref: "#/components/parameters/negocioId"
    get:
      tags: [movimientos]
      summary: Listar movimientos (ventas/servicios) del negocio
      operationId: listMovimientos
      responses:
        "200":
          description: Lista de movimientos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Movimiento"
    post:
      tags: [movimientos]
      summary: Registrar movimiento
      description: >
        Si el producto tiene inventario y la venta lo deja bajo el mínimo
        configurado, responde 409 en vez de aplicar el descuento — el
        frontend debe mostrar el modal de confirmación y reintentar con
        `confirmar_bajo_minimo=true`.
      operationId: createMovimiento
      parameters:
        - name: confirmar_bajo_minimo
          in: query
          required: false
          schema:
            type: boolean
            default: false
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MovimientoCreate"
      responses:
        "201":
          description: Movimiento registrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Movimiento"
        "403":
          description: >
            cliente_vehiculo_id fue enviado pero el negocio no tiene el
            módulo clientes_vehiculos activo.
        "409":
          description: >
            Sin stock disponible, o la venta deja el producto bajo el
            mínimo y falta confirmar_bajo_minimo=true.
  /negocios/{negocioId}/cierres-caja:
    parameters:
      - $ref: "#/components/parameters/negocioId"
    get:
      tags: [cierres-caja]
      summary: Listar cierres de caja
      operationId: listCierresCaja
      responses:
        "200":
          description: Lista de cierres de caja
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CierreCaja"
    post:
      tags: [cierres-caja]
      summary: Cerrar caja para un rango de fechas
      description: >
        Suma los movimientos del negocio dentro de [fecha_inicio, fecha_fin]
        y guarda el desglose (capital/ganancia, efectivo/digital) como un
        registro fijo — no se recalcula si los movimientos cambian después.
      operationId: createCierreCaja
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CierreCajaCreate"
      responses:
        "201":
          description: Cierre de caja creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CierreCaja"
        "400":
          description: fecha_fin anterior a fecha_inicio
  /negocios/{negocioId}/clientes-vehiculos:
    parameters:
      - $ref: "#/components/parameters/negocioId"
    get:
      tags: [clientes-vehiculos]
      summary: Listar clientes/vehículos (requiere módulo activo)
      operationId: listClientesVehiculos
      parameters:
        - name: activo
          in: query
          required: false
          description: Filtra por activos o archivados. Sin este filtro, trae ambos.
          schema:
            type: boolean
      responses:
        "200":
          description: Lista de clientes/vehículos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClienteVehiculo"
        "403":
          description: Módulo no activo para este negocio
    post:
      tags: [clientes-vehiculos]
      summary: Registrar cliente/vehículo para seguimiento
      description: >
        Se llama por separado del registro de la venta/servicio (que no
        pide datos de cliente): esto es el paso opcional de "¿quiere
        seguimiento?" al final de la atención. `fecha_proximo_mantenimiento`
        se calcula en el servidor a partir de `fecha_ultimo_servicio` +
        `intervalo_meses`.
      operationId: createClienteVehiculo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClienteVehiculoCreate"
      responses:
        "201":
          description: Cliente/vehículo creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClienteVehiculo"
        "403":
          description: Módulo no activo para este negocio
  /negocios/{negocioId}/clientes-vehiculos/{clienteVehiculoId}:
    parameters:
      - $ref: "#/components/parameters/negocioId"
      - name: clienteVehiculoId
        in: path
        required: true
        schema:
          type: integer
    patch:
      tags: [clientes-vehiculos]
      summary: Editar cliente/vehículo, o archivarlo con activo=false
      operationId: updateClienteVehiculo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClienteVehiculoUpdate"
      responses:
        "200":
          description: Cliente/vehículo actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClienteVehiculo"
        "403":
          description: Módulo no activo para este negocio
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [clientes-vehiculos]
      summary: Eliminar cliente/vehículo por completo
      description: >
        Distinto de archivar (PATCH activo=false). Si tiene movimientos o
        notificaciones asociadas, responde 409 sugiriendo archivar.
      operationId: deleteClienteVehiculo
      responses:
        "204":
          description: Eliminado
        "403":
          description: Módulo no activo para este negocio
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Tiene movimientos/notificaciones asociadas; archivar en vez de eliminar.
...
components:
  parameters:
    negocioId:
      name: negocioId
      in: path
      required: true
      schema:
        type: integer
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    NotFound:
      description: No encontrado
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
...
  schemas:
...
    MovimientoBase:
      type: object
      properties:
        usuario_id: { type: integer }
        producto_id: { type: integer }
        cliente_vehiculo_id: { type: integer, nullable: true }
        tipo: { type: string, description: "ej. venta, servicio" }
        descripcion: { type: string, nullable: true }
        precio_lista: { type: string, format: decimal }
        precio_final: { type: string, format: decimal }
        metodo_pago: { type: string, nullable: true }
        fecha: { type: string, format: date-time }
      required: [usuario_id, producto_id, tipo]
    Movimiento:
      allOf:
        - $ref: "#/components/schemas/MovimientoBase"
        - type: object
          properties:
            id: { type: integer }
            negocio_id: { type: integer }
          required: [id, negocio_id]
    MovimientoCreate:
      $ref: "#/components/schemas/MovimientoBase"
      description: negocio_id no va en el body — viene del path /negocios/{negocioId}/movimientos.
    CierreCajaCreate:
      type: object
      properties:
        periodo:
          type: string
          description: "Etiqueta libre, ej. diario, semanal, mensual."
        fecha_inicio: { type: string, format: date }
        fecha_fin: { type: string, format: date }
      required: [periodo, fecha_inicio, fecha_fin]
    CierreCaja:
      type: object
      properties:
        id: { type: integer }
        negocio_id: { type: integer }
        periodo: { type: string }
        fecha_inicio: { type: string, format: date }
        fecha_fin: { type: string, format: date }
        total_bruto: { type: string, format: decimal }
        total_capital: { type: string, format: decimal }
        total_ganancia: { type: string, format: decimal }
        total_efectivo: { type: string, format: decimal }
        total_digital: { type: string, format: decimal }
    ClienteVehiculoBase:
      type: object
      properties:
        nombre_cliente: { type: string }
        telefono: { type: string, nullable: true }
        placa: { type: string, nullable: true }
        marca_vehiculo: { type: string, nullable: true }
        modelo_vehiculo: { type: string, nullable: true }
        tipo_aceite: { type: string, nullable: true }
        intervalo_meses: { type: integer, nullable: true }
        fecha_ultimo_servicio: { type: string, format: date, nullable: true }
      required: [nombre_cliente]
    ClienteVehiculoCreate:
      $ref: "#/components/schemas/ClienteVehiculoBase"
    ClienteVehiculoUpdate:
      type: object
      description: Todos los campos opcionales; manda solo los que cambian.
      properties:
        nombre_cliente: { type: string }
        telefono: { type: string, nullable: true }
        placa: { type: string, nullable: true }
        marca_vehiculo: { type: string, nullable: true }
        modelo_vehiculo: { type: string, nullable: true }
        tipo_aceite: { type: string, nullable: true }
        intervalo_meses: { type: integer, nullable: true }
        fecha_ultimo_servicio: { type: string, format: date, nullable: true }
        activo: { type: boolean }
    ClienteVehiculo:
      allOf:
        - $ref: "#/components/schemas/ClienteVehiculoBase"
        - type: object
          properties:
            id: { type: integer }
            negocio_id: { type: integer }
            fecha_proximo_mantenimiento:
              type: string
              format: date
              nullable: true
              description: Calculado por el servidor; no se envía al crear/editar.
            activo: { type: boolean }
          required: [id, negocio_id, activo]
...